requirements.txt
shapely==1.6.3
python-geohash==0.8.5
setuptools==38.2.5